+++ app/serializers/user.js
@@ -1,0 +1,6 @@
+exports.createReponseUser = user => ({
+  id: user.id,
+  firstName: user.firstName,
+  lastName: user.lastName,
+  email: user.email
+});
